accounting/receipts-scanner/requirements.txt
langchain-anthropic
python-dotenv
aiofiles
htmldocx
